inventario-backend/requirements.txt
django==3.2.9
djangorestframework==3.12.4
djangorestframework-jwt==0.1.3
django-cors-headers==3.6.0
djangorestframework-simplejwt==4.6.0
django-rest-passwordreset==1.1.0
django-cleanup==5.1.0
PyJWT==2.0.0
urllib3==1.25.8